Dispathing event from GLFW

+++ src/Hazel/Events/KeyEvent.h
@@ -9,7 +9,7 @@ namespace Hazel
     public:
         inline int GetKeyCode() const { return m_KeyCode; }
 
-        EVENT_CLASS_CATEGORY(EventCategory::Keyboard | EventCategory::Input)
+        EVENT_CLASS_CATEGORY((int)EventCategory::Keyboard | (int)EventCategory::Input)
 
     protected:
         KeyEvent(int keycode) : m_KeyCode(keycode)
